feat: multiple responses same fixed uri
onUriRequested() will now allow for each response to be returned an arbitrary
amount of time (defaulting to infinity, which is the same as current behaviour).
This means that different responses can be returned from the same uri.
This is done by changing the structure of  in order to have a
'response' and 'times' key, the latter being the number of times a response
can be returned. If it's set to -1, that's an infinite amount of reponses
(although count() will still return 1).